+++ packages/graphql/src/schema/pagination.ts
@@ -129,32 +129,3 @@ export function createConnectionWithEdgeProperties({
         },
     };
 }
-
-export function createOffsetLimitStr({
-    offset,
-    limit,
-}: {
-    offset?: number | Integer;
-    limit?: number | Integer;
-}): string {
-    const hasOffset = typeof offset !== "undefined" && offset !== 0;
-    const hasLimit = typeof limit !== "undefined" && limit !== 0;
-    let offsetLimitStr = "";
-
-    if (hasOffset && !hasLimit) {
-        offsetLimitStr = `[${offset}..]`;
-    }
-
-    if (hasLimit && !hasOffset) {
-        offsetLimitStr = `[..${limit}]`;
-    }
-
-    if (hasLimit && hasOffset) {
-        const sliceStart = isNeoInt(offset) ? offset.toNumber() : offset;
-        const itemsToGrab = isNeoInt(limit) ? limit.toNumber() : limit;
-        const sliceEnd = (sliceStart as number) + (itemsToGrab as number);
-        offsetLimitStr = `[${offset}..${sliceEnd}]`;
-    }
-
-    return offsetLimitStr;
-}