+++ packages/graphql/tests/utils/run-cypher.ts
@@ -20,9 +20,9 @@
 import type { Result, Session } from "neo4j-driver";
 
 /** Runs cypher safely, cleaning session afterwars */
-export async function runCypher(session: Session, cypher: string): Promise<Result> {
+export async function runCypher(session: Session, cypher: string, parameters?: any): Promise<Result> {
     try {
-        const result = await session.run(cypher);
+        const result = await session.run(cypher, parameters);
         await session.close();
         return result;
     } catch (err: unknown) {